Move memory related functions to Memory.hpp
