+++ dart/common/StlHelpers.hpp
@@ -34,8 +34,9 @@
 
 #include <cassert>
 #include <cstddef>
-#include <memory>
 #include <vector>
+
+#include "Memory.hpp"
 
 // Macro to suppress -Wunused-parameter and -Wunused-variable warnings in
 // release mode when a variable is only used in assertions.
@@ -43,16 +44,6 @@
 
 namespace dart {
 namespace common {
-
-//==============================================================================
-template<typename T, typename... Args>
-std::unique_ptr<T> make_unique(Args&&... args)
-{
-  return std::unique_ptr<T>(new T(std::forward<Args>(args)...));
-}
-// TODO(JS): This is a stopgap solution as it was omitted from C++11 as "partly
-// an oversight". This can be replaced by std::make_unique<T> of the standard
-// library when we migrate to using C++14.
 
 //==============================================================================
 template <typename T>
@@ -68,4 +59,4 @@ static T getVectorObjectIfAvailable(std::size_t index, const std::vector<T>& vec
 } // namespace common
 } // namespace dart
 
-#endif // DART_COMMON_EMPTY_HPP_
+#endif // DART_COMMON_STLHELPERS_HPP_
