+++ dart/config.hpp.in
@@ -60,6 +60,8 @@
 #cmakedefine01 HAVE_BULLET_COLLISION
 #cmakedefine01 HAVE_FLANN
 
+#cmakedefine01 DART_ENABLE_SIMD
+
 #define DART_ROOT_PATH "@CMAKE_SOURCE_DIR@/"
 #define DART_DATA_PATH "@CMAKE_SOURCE_DIR@/data/"
