+++ dart/math/MathTypes.hpp
@@ -33,11 +33,12 @@
 #define DART_MATH_MATHTYPES_HPP_
 
 #include <map>
-#include <memory>
 #include <vector>
 
 #include <Eigen/Dense>
-#include <Eigen/StdVector>
+
+#include "dart/common/Deprecated.hpp"
+#include "dart/common/Memory.hpp"
 
 //------------------------------------------------------------------------------
 // Types
@@ -58,6 +59,18 @@ inline Vector6d compose(const Eigen::Vector3d& _angular,
 typedef std::vector<Eigen::Vector3d> EIGEN_V_VEC3D;
 typedef std::vector<std::vector<Eigen::Vector3d > > EIGEN_VV_VEC3D;
 
+#if EIGEN_VERSION_AT_LEAST(3,2,1)
+
+template <typename _Tp>
+using aligned_vector = std::vector<_Tp,
+    dart::common::detail::aligned_allocator_cpp11<_Tp>>;
+
+template <typename _Key, typename _Tp, typename _Compare = std::less<_Key>>
+using aligned_map = std::map<_Key, _Tp, _Compare,
+    dart::common::detail::aligned_allocator_cpp11<std::pair<const _Key, _Tp>>>;
+
+#else
+
 template <typename _Tp>
 using aligned_vector = std::vector<_Tp, Eigen::aligned_allocator<_Tp>>;
 
@@ -65,105 +78,15 @@ template <typename _Key, typename _Tp, typename _Compare = std::less<_Key>>
 using aligned_map = std::map<_Key, _Tp, _Compare,
     Eigen::aligned_allocator<std::pair<const _Key, _Tp>>>;
 
-#define DART_DEFINE_CREATE_UNIQUE(class_name)\
-  /** Create unique instance of this class. */\
-  template <typename... Args>\
-  static std::shared_ptr<class_name> createShared(Args&&... args)\
-  {\
-    return ::dart::common::make_unique<class_name>(std::forward<Args>(args)...);\
-  }
-
-#define DART_DEFINE_CREATE_SHARED(class_name)\
-  /** Create shared instance of this class. */\
-  template <typename... Args>\
-  static std::shared_ptr<class_name> createShared(Args&&... args)\
-  {\
-    return std::make_shared<class_name>(std::forward<Args>(args)...);\
-  }
-
-#define DART_DEFINE_CREATE_ALIGNED_SHARED(class_name)\
-  /** Create shared instance of this class. */\
-  template <typename... Args>\
-  static std::shared_ptr<class_name> createShared(Args&&... args)\
-  {\
-    return Eigen::make_aligned_shared<class_name>(std::forward<Args>(args)...);\
-  }
-
-#define DART_DEFINE_CREATE(class_name)\
-  DART_DEFINE_CREATE_SHARED(class_name)\
-  DART_DEFINE_CREATE_UNIQUE(class_name)
-
-#define DART_DEFINE_CREATE_ALIGNED(class_name)\
-  DART_DEFINE_CREATE_ALIGNED_SHARED(class_name)\
-  DART_DEFINE_CREATE_UNIQUE(class_name)
-
-#if EIGEN_VERSION_AT_LEAST(3,2,1)
-
-/// Aligned allocator that is compatible with c++11
-// Ref: https://bitbucket.org/eigen/eigen/commits/f5b7700
-// TODO: Remove this and use Eigen::aligned_allocator once new version of Eigen
-// is released with above commit.
-template <class T>
-class aligned_allocator_cpp11 : public std::allocator<T>
-{
-public:
-  typedef std::size_t     size_type;
-  typedef std::ptrdiff_t  difference_type;
-  typedef T*              pointer;
-  typedef const T*        const_pointer;
-  typedef T&              reference;
-  typedef const T&        const_reference;
-  typedef T               value_type;
-
-  template <class U>
-  struct rebind
-  {
-    typedef aligned_allocator_cpp11<U> other;
-  };
-
-  aligned_allocator_cpp11()
-    : std::allocator<T>() {}
-
-  aligned_allocator_cpp11(const aligned_allocator_cpp11& other)
-    : std::allocator<T>(other) {}
-
-  template <class U>
-  aligned_allocator_cpp11(const aligned_allocator_cpp11<U>& other)
-    : std::allocator<T>(other) {}
-
-  ~aligned_allocator_cpp11() {}
-
-  pointer allocate(size_type num, const void* /*hint*/ = 0)
-  {
-    internal::check_size_for_overflow<T>(num);
-    return static_cast<pointer>( internal::aligned_malloc(num * sizeof(T)) );
-  }
-
-  void deallocate(pointer p, size_type /*num*/)
-  {
-    internal::aligned_free(p);
-  }
-};
-
-template <typename _Tp, typename... _Args>
-inline std::shared_ptr<_Tp> make_aligned_shared(_Args&&... __args)
-{
-  typedef typename std::remove_const<_Tp>::type _Tp_nc;
-  return std::allocate_shared<_Tp>(Eigen::aligned_allocator_cpp11<_Tp_nc>(),
-                                   std::forward<_Args>(__args)...);
-}
-
-#else
-
-template <typename _Tp, typename... _Args>
-inline std::shared_ptr<_Tp> make_aligned_shared(_Args&&... __args)
-{
-  typedef typename std::remove_const<_Tp>::type _Tp_nc;
-  return std::allocate_shared<_Tp>(Eigen::aligned_allocator<_Tp_nc>(),
-                                   std::forward<_Args>(__args)...);
-}
-
 #endif
+
+template <typename _Tp, typename... _Args>
+DART_DEPRECATED(6.2)
+std::shared_ptr<_Tp> make_aligned_shared(_Args&&... __args)
+{
+  return ::dart::common::make_aligned_shared<_Tp, _Args...>(
+        std::forward<_Args>(__args)...);
+}
 
 }  // namespace Eigen
